Fixed delete sauce method so only to user who created it can delete it only

+++ backend/controllers/sauce.js
@@ -23,6 +23,17 @@ exports.createSauce = (req, res, next) => {
 exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
         .then(sauce => {
+            if (!sauce) {
+                return res.status(404).json({
+                    error: new Error('Objet non trouvé !')
+                });
+            }
+            if (sauce.userId !== req.auth.userId) {
+                return res.status(401).json({
+                    error: new Error('Requête non autorisée !')
+                });
+            }
+            // FIXME check sauce.userId is equal to req.auth.userId
             const filename = sauce.imageUrl.split('/images/')[1];
             fs.unlink(`images/${filename}`, () => {
                 Sauce.deleteOne({ _id: req.params.id })
